requirements.txt
requests~=2.32.3
rsa~=4.9
ics~=0.7.2
python-dotenv~=1.0.1
pytz~=2024.2
icalendar~=5.0.13
lxml~=5.3.0